requirements.txt
alembic==1.4.2
amqp==2.6.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
-e git+git@gitlab.studyxiao.cn:studyxiao/haw.git@29436208498da9e3d57aea472c34c5fbf4aceb53#egg=haw
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.10
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.1.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.17
vine==1.3.0
Werkzeug==1.0.1
WTForms==2.3.1
